rename: generateType.ts -> generate.ts로 파일 이름 변경
